sdl: maximize the window instead of going full-screen
This will also hide the border so it takes the entire viewport.

The motivation behind this is that SDL_SetWindowFullscreen has 2 modes:
when SDL_WINDOW_FULLSCREEN is set the viewmode changes, aka the display
resolution, which is bad. I tried using SDL_WINDOW_FULLSCREEN_DESKTOP,
which is supposed to help because it's a "fake" full-screen, as in, it
doesn't resize the display, but it didn't work as expected.